fixing up dashboard
