+++ backend/src/Backend/model/DashboardService.py
@@ -61,72 +61,6 @@ class DashboardService(BaseAPIService):
         except Exception as e:
             self.printError(f"{e}")
             return [False, f"SQL execution error: {e}"]
-    
-    def updateUsername(self, old_username: str, new_username: str) -> Tuple[bool, str]:
-        try:
-            query = """
-            UPDATE "User"
-            SET "userID" = %(new_username)s
-            WHERE "userID" = %(old_username)s;
-            """
-
-            self.print(f"Updating username from {old_username} to {new_username}")
-
-            _, _, error = self._dbTool.executeSQL(query,
-                                                vars={"old_username": old_username, "new_username": new_username},
-                                                commit=True)
-
-            if error:
-                self.printError(error)
-                return [False, str(error)]
-            return [True, "Username updated successfully."]
-        except Exception as e:
-            self.printError(e)
-            return [False, str(e)]
-
-    def updatePassword(self, user_id: str, old_password: str, new_password: str) -> Tuple[bool, str]:
-        try:
-            check_query = """
-            SELECT COUNT(*) FROM "User"
-            WHERE "userID" = %(user_id)s AND "password" = %(old_password)s;
-            """
-
-            connData, cursor, error = self._dbTool.executeSQL(check_query,
-                                                            vars={"user_id": user_id, "old_password": old_password},
-                                                            commit=False, closeConn=False)
-
-            if error:
-                connData.putConn()
-                self.printError(error)
-                return [False, "Password check failed."]
-
-            result = cursor.fetchone()
-            if not result or result[0] == 0:
-                connData.putConn()
-                return [False, "Old password incorrect."]
-
-
-            update_query = """
-            UPDATE "User"
-            SET "password" = %(new_password)s
-            WHERE "userID" = %(user_id)s;
-            """
-
-            _, _, update_error = self._dbTool.executeSQL(update_query,
-                                                        vars={"user_id": user_id, "new_password": new_password},
-                                                        commit=True)
-
-            connData.putConn()
-
-            if update_error:
-                self.printError(update_error)
-                return [False, "Password update failed."]
-            return [True, "Password updated successfully."]
-        except Exception as e:
-            self.printError(e)
-            return [False, str(e)]
-
-    
         
     def getBookingFrequency(self, userId: uuid.UUID, startDateTime: Optional[datetime], endDateTime: Optional[datetime], queryLimit: Optional[int]) -> Tuple[bool, Union[str, List[Dict[str, Any]]]]:
         if (startDateTime is not None and endDateTime is not None and startDateTime > endDateTime):
