+++ frontend/src/view/components/header/Header.jsx
@@ -4,12 +4,12 @@ import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 import IconButton from '@mui/material/IconButton';
 import MenuIcon from '@mui/icons-material/Menu';
-import { useAuth } from "../../../wrappers/AuthContext";
 import { useNavigate } from "react-router-dom";
-import { useState } from 'react';
-import Modal from '@mui/material/Modal';
+import { useRef, useState } from 'react';
 import Button from '@mui/material/Button';
 
+import { useAuth } from "../../../wrappers/AuthContext";
+import { AccountMenu } from '../accountMenu/AccountMenu';
 import "./Header.css"
 
 export function Header() {
@@ -19,19 +19,7 @@ export function Header() {
     const [showProfile, setShowProfile] = useState(false);
     const [showNavBar, setShowNavBar] = useState(false);
 
-    const style = {
-      position: 'absolute',
-      top: '10%',
-      left: '85%',
-      transform: 'translate(-50%, -50%)',
-      width: '25%',
-      bgcolor: 'background.paper',
-      border: '2px solid #000',
-      boxShadow: 24,
-      p: 4,
-    };
-
-    const roles = {1: "User", 2: "Admin", 3: "Manager"}
+    const profileIconRef = useRef(null);
 
     return (
         <Box sx={{ flexGrow: 1 }}>
@@ -44,15 +32,14 @@ export function Header() {
                         aria-label="menu"
                         sx={{ mr: 2 }}
                         onClick = { () => {
-                            console.log("HEHEHEHA")
                             setShowNavBar(!showNavBar)
                           }
                         }
                     >
                         <MenuIcon />
                     </IconButton>
-                    <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-                        News
+                    <Typography variant="h6" component="div" sx={{ flexGrow: 1, fontWeight: "bold" }}>
+                        Room Booking App
                     </Typography>
                     {authUserId !== null && (
                       <>
@@ -67,37 +54,20 @@ export function Header() {
                     {authUserId === null && <Button color="inherit" onClick={() => navigate("/login")}>Login</Button>}
                     {authUserId === null && <Button color="inherit" onClick={() => navigate("/signup")}>Signup</Button>}
 
-                    <Button id="profilePic" onClick={() => setShowProfile(true)}>
+                    <Button id="profilePic" ref={profileIconRef} onClick={() => { setShowProfile(true) }}>
                       <img src={require('../../../resources/profile.png')}></img>
                     </Button>
-                    <Modal
-                      open={showProfile}
-                      onClose={() => setShowProfile(false)}
-                      aria-labelledby="modal-modal-title"
-                      aria-describedby="modal-modal-description"
-                    >
-                      <Box sx={style}>
-                        {username && userPerm && (
-                          <>
-                            <Typography id="modal-modal-description" sx={{ mt: 0 }}>
-                              Username: {username ? username : "Not logged in"}
-                            </Typography>
-                            <Typography id="modal-modal-description" sx={{ mt: 0 }}>
-                              Permission: {""}
-                              {roles[userPerm]}
-                            </Typography>
-                          </>
-                        )}
-                        {(!username || !userPerm) && (
-                          <>
-                            <Typography id="modal-modal-description" sx={{ mt: 0 }}>
-                              Not logged in
-                            </Typography>
-                          </>
-                        )}
-                      </Box>
-                    </Modal>
 
+                    <AccountMenu 
+                      username={username} 
+                      anchorEl={profileIconRef} 
+                      open={showProfile} 
+                      onClose={() => setShowProfile(false) } 
+                      logOutHandler={() => {
+                        setAuthUserId(null)
+                        setUsername(null)
+                        setUserPerm(null)
+                      }}></AccountMenu>
                 </Toolbar>
             </AppBar>
         </Box>
